+++ templates/index.html
@@ -328,6 +328,6 @@
         </div>
     </div>
 
-    <script src="{% static 'js/app.js' %}"></script>
+    <script type="module" src="{% static 'js/app.js' %}"></script>
 </body>
 </html>
